requirements.txt
pydub
pyannote.audio
alive-progress
-i https://download.pytorch.org/whl/cu121 torch 
torchvision
torchaudio
